Add support for audio embed code
