+++ src/components/molecules/player.module.scss
@@ -133,6 +133,9 @@
 .leftButtons > * {
 	margin-right: 24px;
 }
+.leftButtonsCompact > * {
+	margin-right: 8px;
+}
 
 .rightButtons > * {
 	margin-left: 16px;
